True up 'officially supported' LLM models
