+++ cmd/cody-gateway-config/BUILD.bazel
@@ -3,14 +3,13 @@ load("@io_bazel_rules_go//go:def.bzl", "go_binary", "go_library")
 go_library(
     name = "cody-gateway-config_lib",
     srcs = [
-        "dotcom_models.go",
         "main.go",
         "providers.go",
-        "utils.go",
     ],
     importpath = "github.com/sourcegraph/sourcegraph/cmd/cody-gateway-config",
     visibility = ["//visibility:private"],
     deps = [
+        "//cmd/cody-gateway-config/models",
         "//internal/modelconfig",
         "//internal/modelconfig/types",
         "//lib/errors",
